% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConQuR_main_tune_libsize.R
\name{ConQuR_libsize}
\alias{ConQuR_libsize}
\title{Remove batch effects from a taxa read count table (library size is included in the model)}
\usage{
ConQuR_libsize(
  tax_tab,
  batchid,
  covariates,
  libsize_tune = NULL,
  batch_ref,
  logistic_lasso = F,
  quantile_type = "standard",
  simple_match = F,
  lambda_quantile = "2p/n",
  interplt = F,
  delta = 0.4999,
  taus = seq(0.005, 0.995, by = 0.005),
  num_core = 2
)
}
\arguments{
\item{tax_tab}{The taxa read count table, samples (row) by taxa (col).}

\item{batchid}{The batch indicator, must be a factor.}

\item{covariates}{The data.frame contains the key variable of interest and other covariates, e.g., data.frame(key, x1, x2).}

\item{libsize_tune}{The samples' library sizes, which is designed to pass values from \code{Tune_ConQuR_libsize}; do not assign values, keep it as NULL.}

\item{batch_ref}{A character, the name of the reference batch, e.g.,``2''.}

\item{logistic_lasso}{A logical value, TRUE for L1-penalized logistic regression, FALSE for standard logistic regression; default is FALSE.}

\item{quantile_type}{A character, ``standard'' for standard quantile regression, ``lasso'' for L1-penalized quantile regression, ``composite'' for composite quantile regression; default is ``standard''.}

\item{simple_match}{A logical value, TRUE for using the simple quantile-quantile matching, FALSE for not; default is FALSE.}

\item{lambda_quantile}{A character, the penalization parameter in quantile regression if \code{quantile_type}=``lasso'' or ``composite''; only two choices ``2p/n'' or ``2p/logn'', where p is the number of expanded covariates and n is the number of non-zero read count; default is ``2p/n''.}

\item{interplt}{A logical value, TRUE for using the data-driven linear interpolation between zero and non-zero quantiles to stablize border estimates, FALSE for not; default is FALSE.}

\item{delta}{A real constant in (0, 0.5), determing the size of the interpolation window if interplt=TRUE, a larger delta leads to a narrower interpolation window; default is 0.4999.}

\item{taus}{A sequence of quantile levels, determing the ``precision'' of estimating conditional quantile functions; default is seq(0.005, 0.995, by=0.005).}

\item{num_core}{A real constant, the number of cores used for computing; default is 2.}
}
\value{
The corrected taxa read count table, samples (row) by taxa (col).
}
\description{
Remove batch effects from a taxa read count table (library size is included in the model)
}
\details{
\itemize{
  \item Use this version when between-batch library size variability is not considered nuisance and maintained in the corrected taxa read count table.
  \item Choose \code{batch_ref} based on prior knowledge, or try several options, there is no default.
  \item The option ``composite'' of \code{quantile_type} is aggressive, use with caution.
  \item If choose \code{simple_match}=TRUE, \code{logistic_lasso}, \code{quantile_type}, \code{lambda_quantile}, \code{interplt} and \code{delta} won't take effect.
  \item Always use a fine grid of \code{taus} if the size of data is adequate.
}
}
\references{
\itemize{
  \item Ling, W. et al. (2021+). ConQuR: batch effects removal for microbiome data in large-scale epidemiology studies via conditional quantile regression.
  \item Ling, W. et al. (2020+). Statistical inference in quantile regression for zero-inflated outcomes. Statistica Sinica.
  \item Machado, J.A.F., Silva, J.S. (2005). Quantiles for counts. Journal of the American Statistical Association 100(472), 1226–1237.
  \item Koenker, R. & Bassett Jr, G. (1978). Regression quantiles. Econometrica: journal of the Econometric Society, 33-50.
  \item Koenker, R. (2005). Econometric Society Monographs: Quantile Regression. New York: Cambridge University.
  \item Zou, H. & Yuan, M. (2008). Composite quantile regression and the oracle model selection theory. The Annals of Statistics 36, 1108-1126.
}
}
