% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{Plot_PCoA}
\alias{Plot_PCoA}
\title{Stratified PCoA plots}
\usage{
Plot_PCoA(
  TAX,
  factor,
  sub_index = NULL,
  dissimilarity = "Bray",
  GUniFrac_type = "d_0.5",
  tree = NULL,
  main = NULL,
  aa = 1.5
)
}
\arguments{
\item{TAX}{The taxa read count table, samples (row) by taxa (col).}

\item{factor}{The variable for stratification, e.g., batchid or the variable of interest, must be a factor.}

\item{sub_index}{A vector of sample indices, to restrict the analysis to a subgroup of samples, e.g., c(1:5, 15:20); default is NULL.}

\item{dissimilarity}{The dissimilarity type, ``Bray'' for Bray-Curtis dissimilarity, ``Aitch'' for Aitchison dissimilarity, ``GUniFrac'' for generalized UniFrac dissimilarity; default is ``Bray''.}

\item{GUniFrac_type}{The generalized UniFrac type, ``d_1'' for weighted UniFrac, ``d_UW'' for unweighted UniFrac, ``d_VAW'' for variance adjusted weighted UniFrac, ``d_0'' for generalized UniFrac with alpha 0, ``d_0.5'' for generalized UniFrac with alpha 0.5; default is ``d_0.5''.}

\item{tree}{The rooted phylogenetic tree of R class ``phylo'', must be provided when \code{dissimilarity}=``GUniFrac''; default is NULL.}

\item{main}{The title of plot; default is NULL.}

\item{aa}{A real number, the character size for the title.}
}
\value{
Print a PCoA plot.
}
\description{
Stratified PCoA plots
}
\references{
\itemize{
  \item Chen, J., & Chen, M. J. (2018). Package ‘GUniFrac’. The Comprehensive R Archive Network (CRAN).
}
}
