% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{RF_Pred_Regression}
\alias{RF_Pred_Regression}
\title{Predict continuous variables based on a taxa read count table by random forest}
\usage{
RF_Pred_Regression(TAX, variable, fold = 5, seed = 2020)
}
\arguments{
\item{TAX}{The taxa read count table, samples (row) by taxa (col).}

\item{variable}{The continuous variable to predict.}

\item{fold}{The number of folds; default is 5.}

\item{seed}{The seed to generate fold indices for samples; default is 2020.}
}
\value{
A list
\itemize{
  \item pred - A table summarizing the predicted and true values for all samples.
  \item rmse_across_fold - RMSEs across folds.
}
}
\description{
Predict continuous variables based on a taxa read count table by random forest
}
