% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microb_data.R
\docType{data}
\name{Sample_Data}
\alias{Sample_Data}
\title{Example data, a taxa read count table, with batchid, key variable and covariates}
\format{
A taxa read count (273 samples by 100 taxa), batchid and the metadata:
\describe{
  \item{batchid}{factor, with levels 0, 1, 2}
  \item{sbp}{key variable, systolic blood pressure, continuous variable}
  \item{sex}{covariate 1, binary variable}
  \item{race}{covariate 2, binary variable}
  \item{age}{covariate 3, continuous variable}
}
}
\usage{
Sample_Data
}
\description{
A dataset containing 100 taxa from 3 batches, key variable is sbp, with covariates, sex, race and age
}
\keyword{datasets}
