% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConQuR_main_tune_libsize.R
\name{Tune_ConQuR_libsize}
\alias{Tune_ConQuR_libsize}
\title{Tune over variations of ConQuR_libsize}
\usage{
Tune_ConQuR_libsize(
  tax_tab,
  batchid,
  covariates,
  batch_ref_pool,
  logistic_lasso_pool,
  quantile_type_pool,
  simple_match_pool,
  lambda_quantile_pool,
  interplt_pool,
  frequencyL,
  frequencyU,
  cutoff = 0.1,
  delta = 0.4999,
  taus = seq(0.005, 0.995, by = 0.005),
  num_core = 2
)
}
\arguments{
\item{tax_tab}{The taxa read count table, samples (row) by taxa (col).}

\item{batchid}{The batch indicator, must be a factor.}

\item{covariates}{The data.frame contains the key variable of interest and other covariates, e.g., data.frame(key, x1, x2).}

\item{batch_ref_pool}{A vector of characters, the candidates for reference batch, e.g., c(``0'', ``2'').}

\item{logistic_lasso_pool}{A vector of logical values, whether or not using the L1-penalized logistic regression, e.g., c(T, F).}

\item{quantile_type_pool}{A vector of characters, the candidates for quantile regression type, e.g., c(``standard'', ``lasso'').}

\item{simple_match_pool}{A vector of logical values, whether or not using the simple quantile-quantile matching, e.g., c(T, F).}

\item{lambda_quantile_pool}{A vector of characters, the candidates for the penalization parameter in quantile regression (``lasso'' or ``composite''), e.g., c(NA, ``2p/n'', ``2p/logn'').}

\item{interplt_pool}{A vector of logical values, whether or not using the data-driven linear interpolation between zero and non-zero quantiles, e.g., c(T, F).}

\item{frequencyL}{A real constant between 0 and 1, the lower bound of prevalence that needs tuning.}

\item{frequencyU}{A real constant between 0 and 1, the upper bound of prevalence that needs tuning.}

\item{cutoff}{A real constant, the grid size of prevalence for tuning; default is 0.1.}

\item{delta}{A real constant in (0, 0.5), determing the size of the interpolation window if interplt=TRUE, a larger delta leads to a narrower interpolation window; default is 0.4999.}

\item{taus}{A sequence of quantile levels, determing the ``precision'' of estimating conditional quantile functions; default is seq(0.005, 0.995, by=0.005).}

\item{num_core}{A real constant, the number of cores used for computing; default is 2.}
}
\value{
A list
\itemize{
  \item tax_final - The optimal corrected taxa read count table, samples (row) by taxa (col).
  \item method_final - A table summarizing variations of ConQuR_libsize chosen for each prevalence cutoff.
}
}
\description{
Tune over variations of ConQuR_libsize
}
\details{
\itemize{
  \item Use this version when between-batch library size variability is not considered nuisance and maintained in the corrected taxa read count table.
  \item ``original'', i.e., the original data without correction is always a default candidate.
  \item If ``standard'' is one candidate for \code{quantile_type_pool}, always include NA as one candidate for \code{lambda_quantile_pool}.
  \item Be cautious with candidate ``composite'' for \code{quantile_type_pool}, the underlying assumption is strong and the computation might be slow.
  \item The tuning procedure finds the local optimal in each cutoff. If \code{frequencyL}=0.2, \code{frequencyU}=0.5 and \code{cutoff}=0.1, the functions determines the combination achieving maximum removal of batch variations on taxa present in 20\%-30\%, ..., 40\%-50\% of the samples, respectively.
  \item The same reference batch is used across taxa in the final optimal corrected table.
}
}
\references{
\itemize{
  \item Ling, W. et al. (2021+). ConQuR: batch effects removal for microbiome data in large-scale epidemiology studies via conditional quantile regression
  \item Ling, W. et al. (2020+). Statistical inference in quantile regression for zero-inflated outcomes. Statistica Sinica.
  \item Machado, J.A.F., Silva, J.S. (2005). Quantiles for counts. Journal of the American Statistical Association 100(472), 1226–1237.
  \item Koenker, R. & Bassett Jr, G. (1978). Regression quantiles. Econometrica: journal of the Econometric Society, 33-50.
  \item Koenker, R. (2005). Econometric Society Monographs: Quantile Regression. New York: Cambridge University.
  \item Zou, H. & Yuan, M. (2008). Composite quantile regression and the oracle model selection theory. The Annals of Statistics 36, 1108-1126.
  \item Anderson, M. J. (2014). Permutational multivariate analysis of variance (PERMANOVA). Wiley statsref: statistics reference online, 1-15.
}
}
